[REFACTORING] Split IMAPServerIdleSSLCompressTest
